chore(clerk-js): Move cookie related code to core/cookies
