+++ packages/clerk-js/src/core/cookies/safeLock.ts
