--- packages/clerk-js/src/core/services/index.ts
@@ -1,1 +1,0 @@
-export * from './authentication/SessionCookieService';
