+++ packages/clerk-js/src/utils/index.ts
@@ -17,7 +17,6 @@ export * from './props';
 export * from './queryStateParams';
 export * from './querystring';
 export * from './runtime';
-export * from './safeLock';
 export * from './url';
 export * from './web3';
 export * from './windowNavigate';
